--- dearnote/src/main/java/umc/umc/config/SwaggerConfig.java
@@ -1,7 +1,0 @@
-package umc.umc.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-}
