--- dearnote/src/main/java/umc/umc/converter/MemberConverter.java
@@ -1,4 +1,0 @@
-package umc.umc.converter;
-
-public class MemberConverter {
-}
