--- dearnote/src/main/java/umc/umc/domain/Letter.java
@@ -1,23 +1,0 @@
-package umc.umc.domain;
-
-import jakarta.persistence.*;
-import umc.domain.common.BaseEntity;
-import umc.domain.enums.LetterType;
-
-public class Letter extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(15) DEFAULT 'RECEIVED'")
-    private LetterType type;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letterBox_id", nullable = false)
-    private LetterBox letterBox;
-}
