--- dearnote/src/main/java/umc/umc/domain/LetterBox.java
@@ -1,20 +1,0 @@
-package umc.umc.domain;
-
-import jakarta.persistence.*;
-import umc.domain.common.BaseEntity;
-
-public class LetterBox extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @OneToMany(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letter_id", nullable = false)
-    private Letter letter;
-
-}
