--- dearnote/src/main/java/umc/umc/domain/Member.java
@@ -1,26 +1,0 @@
-package umc.umc.domain;
-
-import jakarta.persistence.*;
-import umc.domain.common.BaseEntity;
-import umc.domain.enums.MemberStatus;
-
-public class Member extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(15) DEFAULT 'ACTIVE'")
-    private MemberStatus status;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letterBox_id", nullable = false)
-    private LetterBox letterBox;
-}
