--- dearnote/src/main/java/umc/umc/domain/enums/LetterType.java
@@ -1,5 +1,0 @@
-package umc.umc.domain.enums;
-
-public enum LetterType {
-    RECEIVED, SENT
-}
