--- dearnote/src/main/java/umc/umc/domain/enums/MemberStatus.java
@@ -1,5 +1,0 @@
-package umc.umc.domain.enums;
-
-public enum MemberStatus {
-    ACTIVE, INACTIVE
-}
