--- dearnote/src/main/java/umc/umc/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package umc.umc.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemberRepository {
-}
