--- dearnote/src/main/java/umc/umc/service/MemberQueryService.java
@@ -1,4 +1,0 @@
-package umc.umc.service;
-
-public interface MemberQueryService {
-}
