--- dearnote/src/main/java/umc/umc/web/controller/MemberRestController.java
@@ -1,9 +1,0 @@
-package umc.umc.web.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class MemberRestController {
-}
