--- dearnote/src/main/java/umc/umc/web/dto/MemberRequestDTO.java
@@ -1,4 +1,0 @@
-package umc.umc.web.dto;
-
-public class MemberRequestDTO {
-}
