+++ src/main/java/com/dearnote/config/SwaggerConfig.java
@@ -1,0 +1,7 @@
+package com.dearnote.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+}
