+++ src/main/java/com/dearnote/converter/MemberConverter.java
@@ -1,0 +1,4 @@
+package com.dearnote.converter;
+
+public class MemberConverter {
+}
