+++ src/main/java/com/dearnote/domain/Keyword.java
@@ -1,0 +1,24 @@
+package com.dearnote.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+import com.dearnote.domain.common.BaseEntity;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Keyword extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, columnDefinition = "VARCHAR(100)")
+    private String keyword;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "letter_id", nullable = false)
+    private Letter letter;
+}
