+++ src/main/java/com/dearnote/domain/Member.java
@@ -1,0 +1,36 @@
+package com.dearnote.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+import com.dearnote.domain.common.BaseEntity;
+import com.dearnote.domain.enums.MemberStatus;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Member extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Enumerated(EnumType.STRING)
+    @Column(columnDefinition = "VARCHAR(15) DEFAULT 'ACTIVE'")
+    private MemberStatus status;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "letterBox_id", nullable = false)
+    private LetterBox letterBox;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "letter_id", nullable = false)
+    private Letter letter;
+}
