+++ src/main/java/com/dearnote/domain/common/BaseEntity.java
@@ -1,0 +1,23 @@
+package com.dearnote.domain.common;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+public abstract class BaseEntity {
+
+    @Column(columnDefinition = "DATETIME(6)")
+    private LocalDateTime createdAt;
+
+    @Column(columnDefinition = "DATETIME(6)")
+    private LocalDateTime updatedAt;
+}
