+++ src/main/java/com/dearnote/domain/enums/LetterStatus.java
@@ -1,0 +1,5 @@
+package com.dearnote.domain.enums;
+
+public enum LetterStatus {
+    LOCKED, UNLOCKED
+}
