+++ src/main/java/com/dearnote/domain/enums/LetterType.java
@@ -1,0 +1,5 @@
+package com.dearnote.domain.enums;
+
+public enum LetterType {
+    RECEIVED, SENT
+}
