+++ src/main/java/com/dearnote/domain/enums/MemberStatus.java
@@ -1,0 +1,5 @@
+package com.dearnote.domain.enums;
+
+public enum MemberStatus {
+    ACTIVE, INACTIVE
+}
