+++ src/main/java/com/dearnote/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.dearnote.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class MemberRepository {
+}
