+++ src/main/java/com/dearnote/service/MemberQueryService.java
@@ -1,0 +1,4 @@
+package com.dearnote.service;
+
+public interface MemberQueryService {
+}
