+++ src/main/java/com/dearnote/service/MemberQueryServiceImpl.java
@@ -1,0 +1,9 @@
+package com.dearnote.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberQueryServiceImpl implements MemberQueryService{
+}
