+++ src/main/java/com/dearnote/web/controller/MemberRestController.java
@@ -1,0 +1,9 @@
+package com.dearnote.web.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class MemberRestController {
+}
