+++ src/main/java/com/dearnote/web/dto/MemberRequestDTO.java
@@ -1,0 +1,4 @@
+package com.dearnote.web.dto;
+
+public class MemberRequestDTO {
+}
