--- src/main/java/umc/config/SwaggerConfig.java
@@ -1,7 +1,0 @@
-package umc.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-}
