--- src/main/java/umc/converter/MemberConverter.java
@@ -1,4 +1,0 @@
-package umc.converter;
-
-public class MemberConverter {
-}
