--- src/main/java/umc/domain/Image.java
@@ -1,29 +1,0 @@
-package umc.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc.domain.common.BaseEntity;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Image extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = true, columnDefinition = "VARCHAR(300)")
-    private String originFileName;
-
-    @Column(nullable = true, columnDefinition = "VARCHAR(300)")
-    private String storeFileName;
-
-    private Integer size;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letter_id", nullable = false)
-    private Letter letter;
-}
