--- src/main/java/umc/domain/LetterBox.java
@@ -1,29 +1,0 @@
-package umc.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc.domain.common.BaseEntity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class LetterBox extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @OneToMany(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letter_id", nullable = false)
-    private List<Letter> letterList = new ArrayList<>();
-
-}
