--- src/main/java/umc/domain/Member.java
@@ -1,37 +1,0 @@
-package umc.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc.domain.common.BaseEntity;
-import umc.domain.enums.LetterType;
-import umc.domain.enums.MemberStatus;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Member extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(15) DEFAULT 'ACTIVE'")
-    private MemberStatus status;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letterBox_id", nullable = false)
-    private LetterBox letterBox;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "letter_id", nullable = false)
-    private Letter letter;
-}
