--- src/main/java/umc/domain/enums/LetterStatus.java
@@ -1,5 +1,0 @@
-package umc.domain.enums;
-
-public enum LetterStatus {
-    LOCKED, UNLOCKED
-}
