--- src/main/java/umc/domain/enums/LetterType.java
@@ -1,5 +1,0 @@
-package umc.domain.enums;
-
-public enum LetterType {
-    RECEIVED, SENT
-}
