--- src/main/java/umc/domain/enums/MemberStatus.java
@@ -1,5 +1,0 @@
-package umc.domain.enums;
-
-public enum MemberStatus {
-    ACTIVE, INACTIVE
-}
