--- src/main/java/umc/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package umc.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemberRepository {
-}
