--- src/main/java/umc/service/MemberQueryService.java
@@ -1,4 +1,0 @@
-package umc.service;
-
-public interface MemberQueryService {
-}
