--- src/main/java/umc/service/MemberQueryServiceImpl.java
@@ -1,9 +1,0 @@
-package umc.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberQueryServiceImpl implements MemberQueryService{
-}
