--- src/main/java/umc/web/dto/MemberRequestDTO.java
@@ -1,4 +1,0 @@
-package umc.web.dto;
-
-public class MemberRequestDTO {
-}
