--- src/main/java/umc/web/dto/MemberResponseDTO.java
@@ -1,4 +1,0 @@
-package umc.web.dto;
-
-public class MemberResponseDTO {
-}
